Update README.md
With a passion for fashion here's my very own custom website relating to it.

The template is for the Online Clothing Fashion which includes trending wardrobe collection. The website is made from scratch and doesn't include any copied code. Paro Fashion doesn't include anything related to database, it is just a HTML, CSS and JS template. Though you can find some familiar images used in the website taken from various sources, but the purpose is to only make it more beautiful.